Replace the placeholder shading with the hardware's real colour path
Nothing on this hardware picks an RGB value directly. A polygon names a
palette entry, that entry selects one of 32 ramps per channel in the colour
translate RAM, the pixel's luma indexes along the ramp, and the result goes
through a gamma curve. Textured pixels get their luma from the texel via
the luma RAM; untextured ones use the polygon's own. The previous code
invented a colour from a hash of the texture header, which was only ever a
way to see shapes.

Texturing is now implemented on top of that: texture header decode
(dimensions, origin, wrap and mirror, sheet select), 4-bit texel fetch, and
perspective-correct sampling - u/z, v/z and 1/z interpolate linearly in
screen space, which is exactly what the geometry stage prepared them for.

Three real bugs surfaced while getting there, each of which would have
quietly corrupted the result:

- Texture RAM writes pack two consecutive 32-bit CPU words into one storage
  dword, 16 bits from each (model2_tgp_state::tex0_w). Storing a full word
  per dword left every upper half zero, blanking half of every texture.
- Luma RAM is mapped umask32 0x000000ff, so its index is a dword index, not
  a byte address. It was being indexed by byte, so the luma translation
  window read the wrong entries.
- The CRT sync offsets were hardcoded to the renderer's power-on defaults.
  The game programs them, and reading what it writes moves the projection
  from a guess to the game's own centring (0, 128 rather than 90, -8).

The 3D bitmap is 32-bit now, since the colour path produces 8-bit per
channel and packing it back into RGB555 would throw that away.

Known not right: the scene renders flatter and darker than it should. Only
about an eighth of the 816 polygons per field produce pixels - the rest
project to a point with a very large z, which is consistent with distant
scenery but has not been confirmed. Texture sampling itself is verified
working (a surface samples eight distinct texels across a full 0-63 u
range), so the remaining error is in shading or in the geometry that feeds
it, not in the texture unit.

+++ tests/test_geometry.c
@@ -54,7 +54,7 @@ static uint32_t f2u(float f)
 
 static int count_drawn_pixels(void)
 {
-    const uint16_t *map = geo_get_destmap();
+    const uint32_t *map = geo_get_destmap();
     int n = 0;
     if (!map) return 0;
 
@@ -157,7 +157,7 @@ static const pt2_t TRI_C = { 300.0f, 240.0f };
 
 static bool pixel_set(int x, int y)
 {
-    const uint16_t *map = geo_get_destmap();
+    const uint32_t *map = geo_get_destmap();
     return map && map[(size_t)y * FB_STRIDE + x] != 0;
 }
